+++ examples/ztgateway/variables.tf
@@ -1,0 +1,112 @@
+variable "aws_region" {
+  type        = string
+  description = "The AWS region."
+  default     = "us-west-2"
+}
+
+variable "name_prefix" {
+  type        = string
+  description = "The name prefix for all your resources"
+  default     = "zscc"
+  validation {
+    condition     = length(var.name_prefix) <= 12
+    error_message = "Variable name_prefix must be 12 or less characters."
+  }
+}
+
+variable "vpc_cidr" {
+  type        = string
+  description = "VPC IP CIDR Range. All subnet resources that might get created (public, workload, Zscaler) are derived from this /16 CIDR. If you require creating a VPC smaller than /16, you may need to explicitly define all other subnets via public_subnets, workload_subnets, cc_subnets, and route53_subnets variables"
+  default     = "10.1.0.0/16"
+}
+
+variable "endpoint_subnets" {
+  type        = list(string)
+  description = "Zscaler Subnets to create in VPC. This is only required if you want to override the default subnets that this code creates via vpc_cidr variable."
+  default     = null
+}
+
+variable "az_count" {
+  type        = number
+  description = "Default number of subnets to create based on availability zone"
+  default     = 1
+  validation {
+    condition = (
+      (var.az_count >= 1 && var.az_count <= 3)
+    )
+    error_message = "Input az_count must be set to a single value between 1 and 3. Note* some regions have greater than 3 AZs. Please modify az_count validation in variables.tf if you are utilizing more than 3 AZs in a region that supports it. https://aws.amazon.com/about-aws/global-infrastructure/regions_az/."
+  }
+}
+
+variable "owner_tag" {
+  type        = string
+  description = "populate custom owner tag attribute"
+  default     = "zscc-admin"
+}
+
+variable "exclude_igw" {
+  type        = bool
+  description = "By default, example templates require an Internet Gateway to either be created or already exist. Set this variable to true to ensure this module does not depend on either. Only recommended in niche customer environments where internet egresses through a private connection like Direct Connect or ZT Gateway Service deployments"
+  default     = false
+}
+
+variable "exclude_ngw" {
+  type        = bool
+  description = "By default, example templates require one or more NAT Gateway to either be created or already exist. Set this variable to true to ensure this module does not depend on either. Only recommended in niche customer environments where Cloud Connectors are deployed with Public IP Addresses or ZT Gateway Service deployments"
+  default     = true
+}
+
+variable "byo_endpoint_service_name" {
+  type        = string
+  description = "Exising GWLB Endpoint Service name to associate GWLB Endpoints to. Example string format:  \"com.amazonaws.vpce.<region>.<service id>\""
+  default     = null
+}
+
+variable "az_ids" {
+  type        = list(string)
+  description = <<-EOF
+  By default, this module does a lookup for all regional availability zones marked as available.
+  If creating new Zscaler private subnets, it then automatically loops through in order of the returned list based on the variable az_count.
+  Providing each AWS Zone ID explicitly here will take precedence over var.az_count.
+
+  Example: When deploying a greenfield ZT Gateway template in region us-east-1 and 2 AZs where you want to ensure that new subnets
+  are created in use1-az1 and use1-az5, set this variable to:
+  az_ids = ["use1-az1" "use1-az5"]
+
+  Caution: This argument is not supported in all regions or partitions
+  EOF
+  default     = null
+}
+
+
+# BYO (Bring-your-own) variables list
+
+variable "byo_vpc" {
+  type        = bool
+  description = "Bring your own AWS VPC for Cloud Connector"
+  default     = false
+}
+
+variable "byo_vpc_id" {
+  type        = string
+  description = "User provided existing AWS VPC ID"
+  default     = null
+}
+
+variable "byo_subnets" {
+  type        = bool
+  description = "Bring your own AWS Subnets for Cloud Connector"
+  default     = false
+}
+
+variable "byo_subnet_ids" {
+  type        = list(string)
+  description = "User provided existing AWS Subnet IDs"
+  default     = null
+}
+
+variable "zs_route_table_enabled" {
+  type        = bool
+  description = "For brownfield environments where VPC subnets already exist, set to false to not create a new route table to associate to Zscaler subnet(s). Default is true which means module will try to create new route tables"
+  default     = true
+}